Refactor 079.
